refactor: move Service implementations to services module, and endpoints to new endpoints module
