+++ rust/src/endpoints/ping_request.rs
@@ -9,8 +9,8 @@ mod void;
 
 use crate::flow::file_identifier::{FileIdentifier, IdentifierType, ParsedFileIdentifier};
 use crate::flow::uid::{UID, WLTOKEN};
-use crate::flow::ConnectionHandler;
 use crate::flow::{FlowFuture, FlowMessage, Frame, Result};
+use crate::services::ConnectionHandler;
 use std::net::SocketAddr;
 
 const PING_REQUEST_FILE_IDENTIFIER: ParsedFileIdentifier = ParsedFileIdentifier {
