+++ rust/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod endpoints;
 pub mod fdbserver;
 pub mod flow;
 pub mod services;
